refactor: 컨트롤로 클래스 분리
