--- src/main/java/com/nhnacademy/illuwa/order/controller/OrderHistoryController.java
@@ -1,12 +1,0 @@
-package com.nhnacademy.illuwa.order.controller;
-
-import org.springframework.stereotype.Controller;
-
-
-@Controller
-public class OrderHistoryController {
-    /*@GetMapping("/order_history")
-    public String orderHistory() {
-        return "order/order_history";
-    }*/
-}
