--- src/main/java/com/nhnacademy/illuwa/order/controller/OrderViewController.java
@@ -1,12 +1,0 @@
-package com.nhnacademy.illuwa.order.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class OrderViewController {
-    @GetMapping("/order_view")
-    public String orderView() {
-        return "order/order_view";
-    }
-}
