+++ src/main/java/com/nhnacademy/illuwa/order/controller/member/query/MemberOrderQueryController.java
@@ -1,7 +1,10 @@
-package com.nhnacademy.illuwa.order.controller;
+package com.nhnacademy.illuwa.order.controller.member.query;
 
+
+import com.nhnacademy.illuwa.common.dto.PageResponse;
 import com.nhnacademy.illuwa.order.dto.query.detail.OrderItemResponseDto;
 import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.query.list.OrderListResponse;
 import com.nhnacademy.illuwa.order.service.MemberOrderService;
 import com.nhnacademy.illuwa.review.service.ReviewService;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +12,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.HashMap;
 import java.util.List;
@@ -17,9 +21,25 @@ import java.util.stream.Collectors;
 
 @Controller
 @RequiredArgsConstructor
-public class OrderDetailController {
+public class MemberOrderQueryController {
+
     private final MemberOrderService memberOrderService;
     private final ReviewService reviewService;
+
+    @GetMapping("/order-list")
+    public String orderList(@RequestParam(defaultValue = "0") int page,
+                            @RequestParam(defaultValue = "10") int size,
+                            Model model) {
+
+        PageResponse<OrderListResponse> orderPage = memberOrderService.getOrderHistory(page, size);
+
+        model.addAttribute("orderPage", orderPage);
+        model.addAttribute("currentPage", orderPage.page());
+        model.addAttribute("totalPages", orderPage.totalPages());
+        model.addAttribute("activeMenu", "order-list");
+
+        return "order/order_list";
+    }
 
     @GetMapping("/order-detail/{orderId}")
     public String orderOption(@PathVariable("orderId") Long orderId, Model model) {
@@ -46,4 +66,5 @@ public class OrderDetailController {
 
         return "order/order_detail";
     }
+
 }
