+++ src/main/resources/templates/auth/login.html
@@ -36,7 +36,7 @@
             <a th:href="@{/oauth2/authorization/payco}" class="oauth-button full-width">
                 <img th:src="@{/images/PAYCO_loginButton.png}" alt="PAYCO 로그인">
             </a>
-            <a th:href="@{/guest-login}" class="button button-primary full-width">비회원 주문조회</a>
+            <a th:href="@{/guest-check-order}" class="button button-primary full-width">비회원 주문조회</a>
         </section>
     </div>
 </main>
